added calculation for testing dataset: accuracy, presicion, recall, f-measure. added Confusion Matrix calculation, added ModelTestWorker and signals for it, changed BrushSize in paintView, added slots for testing model page
Former-commit-id: ef00c0ee6e0321cf545c7580b97f87abb2e77b56 [formerly fea83c8c2caaa1cc73844990a56bf32337b03b61]
Former-commit-id: 2b30478c4a1e1ec9f3ac877e4e1ace4e912c4b21

+++ src/core/utils/Sample.hpp
@@ -45,8 +45,8 @@ namespace s21 {
 			}
 		}
 
-		DatasetGroup &operator[](int index)				{ return groups[index]; }
-		const DatasetGroup &operator[](int index) const	{ return groups[index]; }
+		DatasetGroup &operator[](size_t index)				{ return groups[index]; }
+		const DatasetGroup &operator[](size_t index) const	{ return groups[index]; }
 		
 		size_t size()		{ return groups.size(); }
 		bool isUsedUp() 	{ return (size_t) current_iteration >= groups.size(); }
